+++ package.json
@@ -99,28 +99,26 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
-    "qs": "^6.10.1",
-    "history": "^5.0.0",
-    "react-helmet": "^6.1.0",
     "electron-debug": "^3.2.0",
     "electron-log": "^4.4.7",
+    "electron-store": "^8.1.0",
     "electron-updater": "^5.0.3",
+    "history": "^5.0.0",
+    "qs": "^6.10.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
     "react-router-dom": "^6.3.0"
   },
   "devDependencies": {
-    "postcss": "^8.3.6",
-    "tailwindcss": "^2.2.7",
-    "@types/history": "4.7.6",
-    "@types/qs": "^6.9.7",
-    "postcss-loader": "^6.1.1",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.6",
     "@teamsupercell/typings-for-css-modules-loader": "^2.5.1",
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.2.0",
+    "@types/history": "4.7.6",
     "@types/jest": "^27.5.1",
     "@types/node": "17.0.33",
+    "@types/qs": "^6.9.7",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
     "@types/react-test-renderer": "^18.0.0",
@@ -162,6 +160,8 @@
     "jest-environment-jsdom": "^28.1.0",
     "lint-staged": "^12.4.1",
     "mini-css-extract-plugin": "^2.6.0",
+    "postcss": "^8.3.6",
+    "postcss-loader": "^6.1.1",
     "prettier": "^2.6.2",
     "react-refresh": "^0.13.0",
     "react-test-renderer": "^18.1.0",
@@ -169,6 +169,7 @@
     "sass": "^1.51.0",
     "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
+    "tailwindcss": "^2.2.7",
     "terser-webpack-plugin": "^5.3.1",
     "ts-jest": "^28.0.2",
     "ts-loader": "^9.3.0",
