+++ src/renderer/App.tsx
@@ -1,5 +1,9 @@
-import React from 'react';
-import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import {
+  BrowserRouter as Router,
+  Routes,
+  Route,
+  useLocation,
+} from 'react-router-dom';
 import qs from 'qs';
 
 import './App.css';
@@ -7,23 +11,23 @@ import './helpers/fontAwesome';
 import Main from './components/printer/Main';
 import Screen from './components/printer/Screen';
 
+function MyApp() {
+  const location = useLocation();
+  const { search } = location;
+  const so = qs.parse(search.slice(search.lastIndexOf('?') + 1));
+  return so.page === 'screen' ? (
+    <Screen select={so.select?.toString() || 'frame'} />
+  ) : (
+    <Main />
+  );
+}
+
 export default function App() {
   return (
     <Router>
-      <Switch>
-        <Route
-          path="*"
-          render={(props) => {
-            const { search } = props.location;
-            const so = qs.parse(search.slice(search.lastIndexOf('?') + 1));
-            return so.page === 'screen' ? (
-              <Screen select={so.select?.toString() || 'frame'} />
-            ) : (
-              <Main />
-            );
-          }}
-        />
-      </Switch>
+      <Routes>
+        <Route path="*" element={<MyApp />} />
+      </Routes>
     </Router>
   );
 }
