+++ src/renderer/index.tsx
@@ -4,10 +4,3 @@ import App from './App';
 const container = document.getElementById('root')!;
 const root = createRoot(container);
 root.render(<App />);
-
-// calling IPC exposed from preload script
-window.electron.ipcRenderer.once('ipc-example', (arg) => {
-  // eslint-disable-next-line no-console
-  console.log(arg);
-});
-window.electron.ipcRenderer.sendMessage('ipc-example', ['ping']);
