+++ src/renderer/preload.d.ts
@@ -4,12 +4,13 @@ declare global {
   interface Window {
     electron: {
       ipcRenderer: {
-        sendMessage(channel: Channels, args: unknown[]): void;
+        sendMessage(channel: Channels, args: any): void;
         on(
           channel: string,
-          func: (...args: unknown[]) => void
+          func: (...args: any) => void
         ): (() => void) | undefined;
         once(channel: string, func: (...args: unknown[]) => void): void;
+        invoke(channel: Channels, args?: any): any;
       };
     };
   }
